Remove unused functions, move stuff in to cpp/h files
