+++ UnchainedPlugin/UnchainedPlugin.vcxproj.filters
@@ -49,6 +49,9 @@
     <ClCompile Include="Lib\EasyLoggingPP\easylogging++.cc">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="Logging.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="include\main.h">
